requirements.txt
# System
#MySQL-python==1.2.5

# Heroku related
psycopg2==2.7.1
gunicorn==19.7.1
dj-database-url==0.4.2
whitenoise==3.3.0

# Dev
ipdb==0.6.1
ipython==0.12.1
Fabric==1.3.2

# App
Django==1.10
djangorestframework==3.4.4
nose==1.1.2
factory-boy==2.8.1
